Refactor jar_publish
